requirements.txt
pandas
streamlit
st-annotated-text
reportlab